pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "src/signoz_mcp_server/__init__.py"
pattern = "__version__ = ['\"](?P<version>[^'\"]*)['\"]"

[project]
name = "signoz-mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
license = { file = "LICENSE" }
dependencies = [
    "Flask==3.0.0",
    "openai",
    "pytest>=8.4.1",
    "python-dateutil>=2.9.0.post0",
    "pyyaml>=6.0.1",
    "requests>=2.31.0",
    "ruff>=0.12.3",
    "typing-extensions",
    "flaky",
    "pandas",
    "langevals[langevals,openai]",
]

[project.scripts]
signoz-mcp-server = "signoz_mcp_server.mcp_server:main"

[project.optional-dependencies]
prod = [
    "gunicorn",
]

[tool.hatch.build]
exclude = [
    "**/__pycache__",
    "**/*.pyc",
    "**/*.pyo",
    "**/*.pyd",
    "**/.venv",
    "**/env",
    "**/.ruff_cache"
]

[tool.uv]
package = false
